refactor(ci): Move CI service and controller their own package
